Disabled dummy test case in GenerateAssertionsAPITests (this needs to be revised to create a context where raire-java times out).
Seperated AssertionRepositoryTests into three test classes, based on the three active profiles used.
